validate(booking) : fix UI of selected service (#44)

+++ lib/features/order/presentation/widgets/main_detail_ui/price_details.dart
@@ -73,7 +73,7 @@ class PriceDetails extends ConsumerWidget {
           ...order.bookingDetails.map<Widget>((detail) {
             return buildPriceItem(
               detail.name ?? '',
-              formatPrice(detail.price),
+              formatPrice(detail.price.toInt()),
             );
           }),
